Run the test suite with ParallelTestRunner itself
Split `test/runtests.jl` into per-topic files (core, workers, failures,
output, unit, serial, retries) that `runtests(ParallelTestRunner, ARGS)`
discovers and runs in parallel workers. Each file's sandbox gets
ParallelTestRunner, Test and `utils.jl` via `init_code`.

Adjustments needed because the tests now run in worker sandboxes:
- The sample test files used by the "basic use" and "subdir use" tests
  moved to `test/sample_tests/`, which is excluded from discovery and
  passed explicitly, so nested runs don't recurse into the real suite.
- The custom record type is evaluated into `Main` (as on the workers)
  so it deserializes; a bare `eval` would land in the sandbox module.
- `Base.ARGS` in the custom-worker shutdown test replaced with `String[]`.
- `@show_if_error` moved to `utils.jl` so every sandbox has it.
- The "no workers running" check stays in `runtests.jl`, after the outer
  run, so it checks the main process rather than a single worker.
