+++ test/failures.jl
@@ -1,0 +1,215 @@
+@testset "failing test" begin
+    testsuite = Dict(
+        "failing test" => quote
+            println("This test will fail")
+            @test 1 == 2
+        end
+    )
+    error_line = @__LINE__() - 3
+
+    io = IOBuffer()
+    ioc = IOContext(io, :color => true)
+    @test_throws Test.FallbackTestSetException("Test run finished with errors") begin
+        runtests(ParallelTestRunner, ["--verbose"]; testsuite, stdout=ioc, stderr=ioc)
+    end
+
+    str = String(take!(io))
+    @test contains(str, r"failing test.+ failed at")
+    @test contains(str, "$(basename(@__FILE__)):$error_line")
+    @test contains(str, "FAILURE")
+    @test contains(str, "Output generated during execution of '\e[31mfailing test\e[39m':")
+    @test contains(str, "Test Failed")
+    @test contains(str, "1 == 2")
+end
+
+@testset "failed test row printed entirely to stderr" begin
+    testsuite = Dict(
+        "failing test" => quote
+            @test 1 == 2
+        end
+    )
+
+    # use separate streams: the whole row of a failed test must go to stderr
+    out = IOBuffer()
+    err = IOBuffer()
+    @test_throws Test.FallbackTestSetException("Test run finished with errors") begin
+        runtests(ParallelTestRunner, ["--verbose"]; testsuite, stdout=out, stderr=err)
+    end
+
+    outs = String(take!(out))
+    errs = String(take!(err))
+    m = match(r"failing test[^\n]*failed at", errs)
+    @test m !== nothing
+    # in verbose mode the row has three cell separators (time, init time) on
+    # stderr; the init-time cell used to leak to stdout instead
+    @test m !== nothing && count("│", m.match) == 3
+    @test !contains(outs, "failed at")
+end
+
+@testset "nested failure" begin
+    testsuite = Dict(
+        "nested" => quote
+            @test true
+            @testset "foo" begin
+                @test true
+                @testset "bar" begin
+                    @test false
+                end
+            end
+        end
+    )
+    error_line = @__LINE__() - 5
+
+    io = IOBuffer()
+    @test_throws Test.FallbackTestSetException("Test run finished with errors") begin
+        runtests(ParallelTestRunner, ["--verbose"]; testsuite, stdout=io, stderr=io)
+    end
+
+    str = String(take!(io))
+    @test contains(str, r"nested .+ started at")
+    @test contains(str, r"nested .+ failed at")
+    @test contains(str, r"nested .+ \| .+ 2 .+ 1 .+ 3")
+    @test contains(str, r"foo .+ \| .+ 1 .+ 1 .+ 2")
+    @test contains(str, r"bar .+ \| .+ 1 .+ 1")
+    @test contains(str, "FAILURE")
+    @test contains(str, "Error in testset bar")
+    @test contains(str, "$(basename(@__FILE__)):$error_line")
+end
+
+@testset "throwing test" begin
+    testsuite = Dict(
+        "throwing test" => quote
+            error("This test throws an error")
+        end
+    )
+    error_line = @__LINE__() - 3
+
+    io = IOBuffer()
+    @test_throws Test.FallbackTestSetException("Test run finished with errors") begin
+        runtests(ParallelTestRunner, ["--verbose"]; testsuite, stdout=io, stderr=io)
+    end
+
+    str = String(take!(io))
+    @test contains(str, r"throwing test .+ failed at")
+    @test contains(str, "$(basename(@__FILE__)):$error_line")
+    @test contains(str, "FAILURE")
+    @test contains(str, "Error During Test")
+    @test contains(str, "This test throws an error")
+end
+
+@testset "crashing test" begin
+    msg = "This test will crash"
+    testsuite = Dict(
+        "abort" => quote
+            println($(msg))
+            abort() = ccall(:abort, Nothing, ())
+            abort()
+        end
+    )
+
+    io = IOBuffer()
+    ioc = IOContext(io, :color => true)
+    @test_throws Test.FallbackTestSetException("Test run finished with errors") begin
+        runtests(ParallelTestRunner, ["--verbose"]; testsuite, stdout=ioc, stderr=ioc)
+    end
+
+    str = String(take!(io))
+    @test contains(str, "Output generated during execution of '\e[31mabort\e[39m':")
+    # Make sure we can capture the output generated by the crashed process, see
+    # issue <https://github.com/JuliaTesting/ParallelTestRunner.jl/issues/83>.
+    @test contains(str, msg)
+    # "in expression starting at" comes from the abort trap, make sure we
+    # captured that as well.
+    @test contains(str, "in expression starting at")
+    # Following are messages printed by ParallelTestRunner.
+    @test contains(str, r"abort .+ started at")
+    @test contains(str, r"abort.+ crashed at")
+    @test contains(str, "FAILURE")
+    @test contains(str, "Error During Test")
+    @test contains(str, "Malt.TerminatedWorkerException")
+end
+
+@testset "worker task failure detected by monitor" begin
+    testsuite = Dict(
+        "a" => :( @test true ),
+    )
+
+    exception = ErrorException("test_worker exploded")
+    # A bad test_worker will cause the worker task to error out.  With this test
+    # we want to make sure the task monitoring system catches and handles it.
+    test_worker(name) = throw(exception)
+
+    io = IOBuffer()
+    try
+        runtests(ParallelTestRunner, ["--jobs=1"];
+                 test_worker, testsuite, stdout=io, stderr=io)
+        # The `runtests` above should handle the error, so we shouldn't get here
+        @test false
+    catch e
+        @test typeof(e) === TaskFailedException
+        @test first(Base.current_exceptions(e.task)).exception == exception
+    end
+    str = String(take!(io))
+    @test contains(str, "Caught an error, stopping...")
+    @test !contains(str, "SUCCESS")
+    # Not even FAILURE is printed in this case, we exit very early.
+    @test !contains(str, "FAILURE")
+end
+
+@testset "quickfail" begin
+    # To test that `--quickfail` does its job, we want to launch tests with one job, with
+    # the failing one launched as first: we should observe that the successful ones are
+    # never started.
+    testsuite = Dict(
+        "fail-test" => :( @test false ),
+        # If there's a bug in the launch order, having multiple successful tests increases
+        # the likelihood of hitting the bug, making investigation easier.
+        "pass-test1" => :( @test true ),
+        "pass-test2" => :( @test true ),
+        "pass-test3" => :( @test true ),
+        "pass-test4" => :( @test true ),
+        "pass-test5" => :( @test true ),
+    )
+    io = IOBuffer()
+    @test_throws Test.FallbackTestSetException begin
+        # Call `_runtests` so that we can enforce a run order.
+        ParallelTestRunner._runtests(
+            # Use a single job to make sure only `fail-test` is started.
+            ParallelTestRunner, parse_args(["--quickfail", "--verbose", "--jobs=1"]);
+            testsuite,
+            tests=["fail-test", "pass-test1", "pass-test2", "pass-test3", "pass-test4", "pass-test5"],
+            stdout=io,
+            stderr=io,
+        )
+    end
+    str = String(take!(io))
+    @test contains(str, r"fail-test .+ started at")
+    @test contains(str, r"fail-test .+ failed at")
+    @test contains(str, "FAILURE")
+    # `fail-test` is always launched first, `pass-test`s are never started.
+    @test !contains(str, r"pass-test[1-5] .+ started at")
+end
+
+@testset "mixed pass and fail" begin
+    testsuite = Dict(
+        "passes" => quote
+            @test true
+            @test 1 + 1 == 2
+        end,
+        "also_passes" => quote
+            @test true
+        end,
+        "fails" => quote
+            @test false
+        end,
+    )
+    io = IOBuffer()
+    @test_throws Test.FallbackTestSetException begin
+        runtests(ParallelTestRunner, String[]; testsuite, stdout=io, stderr=io)
+    end
+    str = String(take!(io))
+    @test contains(str, "FAILURE")
+    @test contains(str, "passes")
+    @test contains(str, "also_passes")
+    @test contains(str, "fails")
+end
