+++ test/runtests.jl
@@ -1,1717 +1,30 @@
 using ParallelTestRunner
 using Test
 
-# Helper macro to show output of tests in case they fail.  Useful for debugging.
-macro show_if_error(io, expr)
-    quote
-        try
-            @elapsed $(esc(expr))
-        catch
-            output = String(take!($(esc(io))))
-            printstyled(stderr, "Output of failed test >>>>>>>>>>>>>>>>>>>>\n", color=:red, bold=true)
-            println(stderr, output)
-            printstyled(stderr, "End of output <<<<<<<<<<<<<<<<<<<<<<<<<<<<\n", color=:red, bold=true)
-            rethrow()
-        end
-    end
-end
-
-cd(@__DIR__)
-
 include(joinpath(@__DIR__, "utils.jl"))
 
-@testset "ParallelTestRunner" verbose=true begin
-
-@testset "basic use" begin
-    io = IOBuffer()
-    io_color = IOContext(io, :color => true)
-    runtests(ParallelTestRunner, ["--verbose"]; stdout=io_color, stderr=io_color)
-    str = String(take!(io))
-
-    println()
-    println("Showing the output of one test run:")
-    println("-"^80)
-    print(str)
-    println("-"^80)
-    println()
-
-    @test contains(str, "SUCCESS")
-
-    # --verbose output
-    @test contains(str, r"basic .+ started at")
-
-    @test contains(str, "time (s)")
-
-    @test contains(str, "Available memory:")
-    @test contains(str, "Max worker RSS:")
-    @test contains(str, "Init")
-
-    # compile time as part of the struct not available before 1.11
-    if VERSION >= v"1.11"
-        @test contains(str, "Compile")
-        @test contains(str, "(%)")
-    end
-
-    @test isfile(ParallelTestRunner.get_history_file(ParallelTestRunner))
+# Every test file exercises `runtests` itself, so each one is run in its own worker by
+# ParallelTestRunner: the nested runs spawn their own workers from there.
+testsuite = find_tests(@__DIR__)
+filter!(testsuite) do (name, _)
+    # `utils.jl` is a helper, and `sample_tests/` holds test files used by the tests themselves.
+    name != "utils" && !startswith(name, "sample_tests/")
 end
 
-@testset "default njobs" begin
-    @test ParallelTestRunner.default_njobs(; cpu_threads=4, free_memory=UInt64(2) ^ 28) == 1
-    @test ParallelTestRunner.default_njobs(; cpu_threads=4, free_memory=UInt64(2) ^ 30) == 1
-    @test ParallelTestRunner.default_njobs(; cpu_threads=4, free_memory=UInt64(2) ^ 31) == 1
-    @test ParallelTestRunner.default_njobs(; cpu_threads=4, free_memory=UInt64(2) ^ 32) == 2
-    @test ParallelTestRunner.default_njobs(; cpu_threads=4, free_memory=UInt64(2) ^ 33) == 4
-    @test ParallelTestRunner.default_njobs(; cpu_threads=4, free_memory=UInt64(2) ^ 34) == 4
+init_code = quote
+    using ParallelTestRunner
+    using Test
+    include($(joinpath(@__DIR__, "utils.jl")))
 end
 
-@testset "subdir use" begin
-    d = @__DIR__
-    testsuite = find_tests(d)
-    @test last(testsuite["basic"].args) == joinpath(d, "basic.jl")
-    @test last(testsuite["subdir/subdir_test"].args) == joinpath(d, "subdir", "subdir_test.jl")
-end
+push!(ARGS, "--verbose")
 
-@testset "custom tests" begin
-    testsuite = Dict(
-        "custom" => quote
-            @test true
-        end
-    )
+runtests(ParallelTestRunner, ARGS; testsuite, init_code)
 
-    io = IOBuffer()
-    runtests(ParallelTestRunner, ["--verbose"]; testsuite, stdout=io, stderr=io)
-
-    str = String(take!(io))
-    @test !contains(str, r"basic .+ started at")
-    @test contains(str, r"custom .+ started at")
-    @test contains(str, "SUCCESS")
-end
-
-@testset "test start order" begin
-    # Tests must start in the given order, regardless of the
-    # number of threads of the host Julia process (issue #139).
-    testsuite = Dict(
-        "sort$(n)" => :(@test true)
-        for n in 1:5
-    )
-    tests = ["sort$(n)" for n in 1:length(testsuite)]
-
-    io = IOBuffer()
-    # with a single job, tests start strictly in acquisition order
-    ParallelTestRunner._runtests(
-        ParallelTestRunner, parse_args(["--jobs=1", "--verbose"]);
-        testsuite,
-        tests,
-        stdout=io,
-        stderr=io,
-    )
-
-    str = String(take!(io))
-    @test contains(str, "SUCCESS")
-    started_at = map(tests) do name
-        m = match(Regex("$(name) .+ started at"), str)
-        @test m !== nothing
-        m === nothing ? typemax(Int) : m.offset
-    end
-    @test issorted(started_at)
-end
-
-@testset "init code" begin
-    init_code = quote
-        using Test
-        should_be_defined() = true
-
-        macro should_also_be_defined()
-            return :(true)
-        end
-    end
-    testsuite = Dict(
-        "custom" => quote
-            @test should_be_defined()
-            @test @should_also_be_defined()
-        end
-    )
-
-    io = IOBuffer()
-    runtests(ParallelTestRunner, ["--verbose"]; init_code, testsuite, stdout=io, stderr=io)
-
-    str = String(take!(io))
-    @test contains(str, r"custom .+ started at")
-    @test contains(str, "SUCCESS")
-end
-
-@testset "init worker code" begin
-    init_worker_code = quote
-        should_be_defined() = true
-
-        macro should_also_be_defined()
-            return :(true)
-        end
-    end
-    init_code = quote
-        using Test
-        import ..should_be_defined, ..@should_also_be_defined
-    end
-
-    testsuite = Dict(
-        "custom" => quote
-            @test should_be_defined()
-            @test @should_also_be_defined()
-        end
-    )
-
-    io = IOBuffer()
-    runtests(ParallelTestRunner, ["--verbose"]; init_code, init_worker_code, testsuite, stdout=io, stderr=io)
-
-    str = String(take!(io))
-    @test contains(str, r"custom .+ started at")
-    @test contains(str, "SUCCESS")
-end
-
-@testset "custom record type" begin
-    # Custom record wraps the default `TestRecord` and adds one field. It must
-    # be defined on both the main process and every worker (via
-    # init_worker_code) because the record crosses the Malt serialization
-    # boundary.
-    init_worker_code = quote
-        using ParallelTestRunner: TestRecord
-        struct MyRecord <: ParallelTestRunner.AbstractTestRecord
-            base::TestRecord
-            extra::String
-        end
-        function ParallelTestRunner.execute(
-            ::Type{MyRecord}, mod::Module, f, name, start_time, custom_args,
-            )
-            base = ParallelTestRunner.execute(TestRecord, mod, f, name, start_time, custom_args)
-            MyRecord(base, custom_args.tag)
-        end
-        function ParallelTestRunner.print_test_finished(
-            record::MyRecord, wrkr, test, ctx::ParallelTestRunner.TestIOContext,
-            )
-            lock(ctx.lock)
-            try
-                println(ctx.stdout, "EXTRA[$test]=$(record.extra)")
-                flush(ctx.stdout)
-            finally
-                unlock(ctx.lock)
-            end
-        end
-    end
-
-    eval(init_worker_code)  # also define on the main process so the record deserializes
-
-    testsuite = Dict(
-        "custom" => quote
-            @test 1 + 1 == 2
-        end,
-    )
-
-    io = IOBuffer()
-    runtests(ParallelTestRunner, ["--verbose"]; testsuite,
-             init_worker_code, RecordType = MyRecord,
-             custom_args = (; tag = "hello"),
-             stdout = io, stderr = io)
-    str = String(take!(io))
-
-    @test contains(str, "EXTRA[custom]=hello")
-    @test contains(str, "SUCCESS")
-end
-
-@testset "custom worker" begin
-    function test_worker(name)
-        if name == "needs env var"
-            return addworker(env = ["SPECIAL_ENV_VAR" => "42"])
-        elseif name == "threads/2"
-            return addworker(exeflags = ["--threads=2"])
-        end
-        return nothing
-    end
-    testsuite = Dict(
-        "needs env var" => quote
-            @test ENV["SPECIAL_ENV_VAR"] == "42"
-        end,
-        "doesn't need env var" => quote
-            @test !haskey(ENV, "SPECIAL_ENV_VAR")
-        end,
-        "threads/1" => quote
-            @test Base.Threads.nthreads() == 1
-        end,
-        "threads/2" => quote
-            @test Base.Threads.nthreads() == 2
-        end
-    )
-
-    io = IOBuffer()
-    runtests(ParallelTestRunner, ["--verbose"]; test_worker, testsuite, stdout=io, stderr=io)
-
-    str = String(take!(io))
-    @test contains(str, r"needs env var .+ started at")
-    @test contains(str, r"doesn't need env var .+ started at")
-    @test contains(str, r"threads/1 .+ started at")
-    @test contains(str, r"threads/2 .+ started at")
-    @test contains(str, "SUCCESS")
-end
-
-@testset "global worker kwargs" begin
-    # `exename`/`exeflags`/`env` on runtests should propagate to every
-    # default-pool worker. We verify via an environment variable propagated
-    # through `env`, Julia flags threaded through `exeflags`, and `exename`
-    # supplied as a `Cmd` prefixing the julia binary (what CUDA.jl uses to
-    # wrap julia with `compute-sanitizer`).
-    testsuite = Dict(
-        "env var" => quote
-            @test ENV["GLOBAL_WORKER_TEST"] == "yes"
-        end,
-        "threads" => quote
-            @test Base.Threads.nthreads() == 2
-        end,
-    )
-    io = IOBuffer()
-    runtests(ParallelTestRunner, ["--verbose"]; testsuite,
-             env = ["GLOBAL_WORKER_TEST" => "yes"],
-             exeflags = ["--threads=2"],
-             exename = `$(Base.julia_cmd()[1])`,  # trivial Cmd wrapping julia
-             stdout = io, stderr = io)
-    str = String(take!(io))
-    @test contains(str, "SUCCESS")
-end
-
-@testset "custom worker with `init_worker_code`" begin
-    init_worker_code = quote
-        should_be_defined() = true
-    end
-    init_code = quote
-        using Test
-        import ..should_be_defined
-    end
-    function test_worker(name, init_worker_code)
-        if name == "needs env var"
-            return addworker(env = ["SPECIAL_ENV_VAR" => "42"]; init_worker_code)
-        elseif name == "threads/2"
-            return addworker(exeflags = ["--threads=2"]; init_worker_code)
-        end
-        return nothing
-    end
-    testsuite = Dict(
-        "needs env var" => quote
-            @test ENV["SPECIAL_ENV_VAR"] == "42"
-            @test should_be_defined()
-        end,
-        "doesn't need env var" => quote
-            @test !haskey(ENV, "SPECIAL_ENV_VAR")
-            @test should_be_defined()
-        end,
-        "threads/1" => quote
-            @test Base.Threads.nthreads() == 1
-            @test should_be_defined()
-        end,
-        "threads/2" => quote
-            @test Base.Threads.nthreads() == 2
-            @test should_be_defined()
-        end
-    )
-
-    io = IOBuffer()
-    runtests(ParallelTestRunner, ["--verbose"]; test_worker, init_code, init_worker_code, testsuite, stdout=io, stderr=io)
-
-    str = String(take!(io))
-    @test contains(str, r"needs env var .+ started at")
-    @test contains(str, r"doesn't need env var .+ started at")
-    @test contains(str, r"threads/1 .+ started at")
-    @test contains(str, r"threads/2 .+ started at")
-    @test contains(str, "SUCCESS")
-end
-
-@testset "failing test" begin
-    testsuite = Dict(
-        "failing test" => quote
-            println("This test will fail")
-            @test 1 == 2
-        end
-    )
-    error_line = @__LINE__() - 3
-
-    io = IOBuffer()
-    ioc = IOContext(io, :color => true)
-    @test_throws Test.FallbackTestSetException("Test run finished with errors") begin
-        runtests(ParallelTestRunner, ["--verbose"]; testsuite, stdout=ioc, stderr=ioc)
-    end
-
-    str = String(take!(io))
-    @test contains(str, r"failing test.+ failed at")
-    @test contains(str, "$(basename(@__FILE__)):$error_line")
-    @test contains(str, "FAILURE")
-    @test contains(str, "Output generated during execution of '\e[31mfailing test\e[39m':")
-    @test contains(str, "Test Failed")
-    @test contains(str, "1 == 2")
-end
-
-@testset "failed test row printed entirely to stderr" begin
-    testsuite = Dict(
-        "failing test" => quote
-            @test 1 == 2
-        end
-    )
-
-    # use separate streams: the whole row of a failed test must go to stderr
-    out = IOBuffer()
-    err = IOBuffer()
-    @test_throws Test.FallbackTestSetException("Test run finished with errors") begin
-        runtests(ParallelTestRunner, ["--verbose"]; testsuite, stdout=out, stderr=err)
-    end
-
-    outs = String(take!(out))
-    errs = String(take!(err))
-    m = match(r"failing test[^\n]*failed at", errs)
-    @test m !== nothing
-    # in verbose mode the row has three cell separators (time, init time) on
-    # stderr; the init-time cell used to leak to stdout instead
-    @test m !== nothing && count("│", m.match) == 3
-    @test !contains(outs, "failed at")
-end
-
-@testset "nested failure" begin
-    testsuite = Dict(
-        "nested" => quote
-            @test true
-            @testset "foo" begin
-                @test true
-                @testset "bar" begin
-                    @test false
-                end
-            end
-        end
-    )
-    error_line = @__LINE__() - 5
-
-    io = IOBuffer()
-    @test_throws Test.FallbackTestSetException("Test run finished with errors") begin
-        runtests(ParallelTestRunner, ["--verbose"]; testsuite, stdout=io, stderr=io)
-    end
-
-    str = String(take!(io))
-    @test contains(str, r"nested .+ started at")
-    @test contains(str, r"nested .+ failed at")
-    @test contains(str, r"nested .+ \| .+ 2 .+ 1 .+ 3")
-    @test contains(str, r"foo .+ \| .+ 1 .+ 1 .+ 2")
-    @test contains(str, r"bar .+ \| .+ 1 .+ 1")
-    @test contains(str, "FAILURE")
-    @test contains(str, "Error in testset bar")
-    @test contains(str, "$(basename(@__FILE__)):$error_line")
-end
-
-@testset "throwing test" begin
-    testsuite = Dict(
-        "throwing test" => quote
-            error("This test throws an error")
-        end
-    )
-    error_line = @__LINE__() - 3
-
-    io = IOBuffer()
-    @test_throws Test.FallbackTestSetException("Test run finished with errors") begin
-        runtests(ParallelTestRunner, ["--verbose"]; testsuite, stdout=io, stderr=io)
-    end
-
-    str = String(take!(io))
-    @test contains(str, r"throwing test .+ failed at")
-    @test contains(str, "$(basename(@__FILE__)):$error_line")
-    @test contains(str, "FAILURE")
-    @test contains(str, "Error During Test")
-    @test contains(str, "This test throws an error")
-end
-
-@testset "crashing test" begin
-    msg = "This test will crash"
-    testsuite = Dict(
-        "abort" => quote
-            println($(msg))
-            abort() = ccall(:abort, Nothing, ())
-            abort()
-        end
-    )
-
-    io = IOBuffer()
-    ioc = IOContext(io, :color => true)
-    @test_throws Test.FallbackTestSetException("Test run finished with errors") begin
-        runtests(ParallelTestRunner, ["--verbose"]; testsuite, stdout=ioc, stderr=ioc)
-    end
-
-    str = String(take!(io))
-    @test contains(str, "Output generated during execution of '\e[31mabort\e[39m':")
-    # Make sure we can capture the output generated by the crashed process, see
-    # issue <https://github.com/JuliaTesting/ParallelTestRunner.jl/issues/83>.
-    @test contains(str, msg)
-    # "in expression starting at" comes from the abort trap, make sure we
-    # captured that as well.
-    @test contains(str, "in expression starting at")
-    # Following are messages printed by ParallelTestRunner.
-    @test contains(str, r"abort .+ started at")
-    @test contains(str, r"abort.+ crashed at")
-    @test contains(str, "FAILURE")
-    @test contains(str, "Error During Test")
-    @test contains(str, "Malt.TerminatedWorkerException")
-end
-
-@testset "worker task failure detected by monitor" begin
-    testsuite = Dict(
-        "a" => :( @test true ),
-    )
-
-    exception = ErrorException("test_worker exploded")
-    # A bad test_worker will cause the worker task to error out.  With this test
-    # we want to make sure the task monitoring system catches and handles it.
-    test_worker(name) = throw(exception)
-
-    io = IOBuffer()
-    try
-        runtests(ParallelTestRunner, ["--jobs=1"];
-                 test_worker, testsuite, stdout=io, stderr=io)
-        # The `runtests` above should handle the error, so we shouldn't get here
-        @test false
-    catch e
-        @test typeof(e) === TaskFailedException
-        @test first(Base.current_exceptions(e.task)).exception == exception
-    end
-    str = String(take!(io))
-    @test contains(str, "Caught an error, stopping...")
-    @test !contains(str, "SUCCESS")
-    # Not even FAILURE is printed in this case, we exit very early.
-    @test !contains(str, "FAILURE")
-end
-
-@testset "test output" begin
-    msg = "This is some output from the test"
-    testsuite = Dict(
-        "output" => quote
-            println($(msg))
-        end
-    )
-
-    io = IOBuffer()
-    runtests(ParallelTestRunner, ["--verbose"]; testsuite, stdout=io, stderr=io)
-
-    str = String(take!(io))
-    @test contains(str, r"output .+ started at")
-    @test contains(str, msg)
-    @test contains(str, "SUCCESS")
-
-    msg2 = "More output"
-    testsuite = Dict(
-        "verbose-1" => quote
-            print($(msg))
-        end,
-        "verbose-2" => quote
-            println($(msg2))
-        end,
-        "silent" => quote
-            @test true
-        end,
-    )
-    io = IOBuffer()
-    # Run all tests on the same worker, makre sure all the output is captured
-    # and attributed to the correct test set.
-    runtests(ParallelTestRunner, ["--verbose", "--jobs=1"]; testsuite, stdout=io, stderr=io)
-
-    str = String(take!(io))
-    @test contains(str, r"verbose-1 .+ started at")
-    @test contains(str, r"verbose-2 .+ started at")
-    @test contains(str, r"silent .+ started at")
-    @test contains(str, "Output generated during execution of 'verbose-1':\n[ $(msg)")
-    @test contains(str, "Output generated during execution of 'verbose-2':\n[ $(msg2)")
-    @test !contains(str, "Output generated during execution of 'silent':")
-    @test contains(str, "SUCCESS")
-end
-
-@testset "warnings" begin
-    testsuite = Dict(
-        "warning" => quote
-            @test_warn "3.0" @warn "3.0"
-        end
-    )
-
-    io = IOBuffer()
-    runtests(ParallelTestRunner, ["--verbose"]; testsuite, stdout=io, stderr=io)
-
-    str = String(take!(io))
-    @test contains(str, r"warning .+ started at")
-    @test contains(str, "SUCCESS")
-end
-
-# Issue <https://github.com/JuliaTesting/ParallelTestRunner.jl/issues/69>.
-@testset "colorful output" begin
-    testsuite = Dict(
-        "color" => quote
-            printstyled("Roses Are Red"; color=:red)
-        end
-    )
-    io = IOBuffer()
-    ioc = IOContext(io, :color => true)
-    runtests(ParallelTestRunner, String[]; testsuite, stdout=ioc, stderr=ioc)
-    str = String(take!(io))
-    @test contains(str, "\e[31mRoses Are Red\e[39m\n")
-    @test contains(str, "SUCCESS")
-
-    testsuite = Dict(
-        "no color" => quote
-            print("Violets are ")
-            printstyled("blue"; color=:blue)
-        end
-    )
-    io = IOBuffer()
-    ioc = IOContext(io, :color => false)
-    runtests(ParallelTestRunner, String[]; testsuite, stdout=ioc, stderr=ioc)
-    str = String(take!(io))
-    @test contains(str, "Violets are blue\n")
-    @test contains(str, "SUCCESS")
-end
-
-@testset "reuse of workers" begin
-    testsuite = Dict(
-        "a" => :(),
-        "b" => :(),
-        "c" => :(),
-        "d" => :(),
-        "e" => :(),
-        "f" => :(),
-    )
-    io = IOBuffer()
-    ioc = IOContext(io, :color => true)
-    old_id_counter = ParallelTestRunner.ID_COUNTER[]
-    njobs = 1
-    runtests(ParallelTestRunner, ["--jobs=$(njobs)"]; testsuite, stdout=ioc, stderr=ioc)
-    str = String(take!(io))
-    @test contains(str, "Running $(length(testsuite)) tests using $(njobs) parallel jobs")
-    @test ParallelTestRunner.ID_COUNTER[] == old_id_counter + njobs
-end
-
-
-# Issue <https://github.com/JuliaTesting/ParallelTestRunner.jl/issues/106>.
-@testset "default workers stopped at end" begin
-    # Use default workers (no test_worker) so the framework creates and should stop them.
-    # More tests than workers so some tasks finish early and must stop their worker.
-    testsuite = Dict(
-        "t1" => :(),
-        "t2" => :(),
-        "t3" => :(),
-        "t4" => :(),
-        "t5" => :(),
-        "t6" => quote
-            # Make this test run longer than the others so that it runs alone...
-            sleep(5)
-            children = _count_child_pids($(getpid()))
-            # ...then check there's only one worker still running. WARNING: this test may be
-            # flaky on very busy systems, if at this point some of the other tests are still
-            # running, hope for the best.
-            if children >= 0
-                @test children == 1
-            end
-        end,
-    )
-    before = _count_child_pids()
-    if before < 0
-        # Counting child PIDs not supported on this platform
-        @test_skip false
-    else
-        old_id_counter = ParallelTestRunner.ID_COUNTER[]
-        njobs = 2
-        io = IOBuffer()
-        ioc = IOContext(io, :color => true)
-        @show_if_error io runtests(ParallelTestRunner, ["--jobs=$(njobs)", "--verbose"];
-                                   testsuite, stdout=ioc, stderr=ioc, init_code=:(include($(joinpath(@__DIR__, "utils.jl")))))
-        # Make sure we didn't spawn more workers than expected.
-        @test ParallelTestRunner.ID_COUNTER[] == old_id_counter + njobs
-        # Allow a moment for worker processes to exit
-        for _ in 1:50
-            sleep(0.1)
-            after = _count_child_pids()
-            after >= 0 && after <= before && break
-        end
-        after = _count_child_pids()
-        @test after >= 0
-        @test after == before
-    end
-end
-
-# Custom workers are handled differently:
-# <https://github.com/JuliaTesting/ParallelTestRunner.jl/pull/107#issuecomment-3980645143>.
-# But we still want to make sure they're terminated at the end.
-@testset "custom workers stopped at end" begin
-    testsuite = Dict(
-        "a" => :(),
-        "b" => :(),
-        "c" => :(),
-        "d" => :(),
-        "e" => :(),
-        "f" => :(),
-    )
-    procs = Base.Process[]
-    procs_lock = ReentrantLock()
-    function test_worker(name)
-        wrkr = addworker()
-        Base.@lock procs_lock push!(procs, wrkr.w.proc)
-        return wrkr
-    end
-    runtests(ParallelTestRunner, Base.ARGS; test_worker, testsuite, stdout=devnull, stderr=devnull)
-    @test all(!Base.process_running, procs)
-end
-
-# ── Unit tests for internal helpers ──────────────────────────────────────────
-
-@testset "extract_flag!" begin
-    args = ["--verbose", "--jobs=4", "test1"]
-    result = ParallelTestRunner.extract_flag!(args, "--verbose")
-    @test result === Some(nothing)
-    @test args == ["--jobs=4", "test1"]
-
-    args = ["--verbose", "--jobs=4", "test1"]
-    result = ParallelTestRunner.extract_flag!(args, "--jobs"; typ=Int)
-    @test something(result) == 4
-    @test args == ["--verbose", "test1"]
-
-    args = ["--verbose", "test1"]
-    result = ParallelTestRunner.extract_flag!(args, "--jobs")
-    @test result === nothing
-    @test args == ["--verbose", "test1"]
-
-    args = ["--format=json"]
-    result = ParallelTestRunner.extract_flag!(args, "--format")
-    @test something(result) == "json"
-    @test isempty(args)
-
-    # values containing `=` are preserved
-    args = ["--format=key=value"]
-    result = ParallelTestRunner.extract_flag!(args, "--format")
-    @test something(result) == "key=value"
-    @test isempty(args)
-
-    # flags sharing a prefix are not captured
-    args = ["--listing"]
-    result = ParallelTestRunner.extract_flag!(args, "--list")
-    @test result === nothing
-    @test args == ["--listing"]
-
-    # a typed flag without a value is an error
-    args = ["--jobs"]
-    @test_throws ErrorException ParallelTestRunner.extract_flag!(args, "--jobs"; typ=Int)
-
-    # a typed flag with an unparseable value is an error
-    args = ["--jobs=abc"]
-    @test_throws ErrorException ParallelTestRunner.extract_flag!(args, "--jobs"; typ=Int)
-end
-
-@testset "parse_args" begin
-    @testset "individual flags" begin
-        args = parse_args(["--verbose"])
-        @test args.verbose !== nothing
-        @test args.jobs === nothing
-        @test args.quickfail === nothing
-        @test args.list === nothing
-        @test isempty(args.positionals)
-
-        args = parse_args(["--jobs=4"])
-        @test something(args.jobs) == 4
-        @test args.verbose === nothing
-
-        args = parse_args(["--quickfail"])
-        @test args.quickfail !== nothing
-        @test args.verbose === nothing
-
-        args = parse_args(["--list"])
-        @test args.list !== nothing
-    end
-
-    @testset "combined flags" begin
-        args = parse_args(["--verbose", "--quickfail", "--jobs=2"])
-        @test args.verbose !== nothing
-        @test args.quickfail !== nothing
-        @test something(args.jobs) == 2
-    end
-
-    @testset "positional arguments" begin
-        args = parse_args(["--verbose", "basic", "subdir"])
-        @test args.verbose !== nothing
-        @test args.positionals == ["basic", "subdir"]
-
-        args = parse_args(["test1", "test2"])
-        @test args.positionals == ["test1", "test2"]
-    end
-
-    @testset "custom arguments" begin
-        args = parse_args(["--gpu", "--nocuda"]; custom=["gpu", "nocuda", "other"])
-        @test args.custom["gpu"] !== nothing
-        @test args.custom["nocuda"] !== nothing
-        @test args.custom["other"] === nothing
-    end
-
-    @testset "unknown flags" begin
-        @test_throws ErrorException parse_args(["--unknown-flag"])
-        @test_throws ErrorException parse_args(["--verbose", "--bogus"])
-    end
-
-    @testset "malformed and lookalike flags" begin
-        # prefix lookalikes are unknown options, not misparsed builtin flags
-        @test_throws ErrorException parse_args(["--verbose-foo"])
-        @test_throws ErrorException parse_args(["--jobs2=3"])
-
-        # a custom flag sharing a builtin's prefix is not captured by the builtin
-        args = parse_args(["--listing"]; custom=["listing"])
-        @test args.list === nothing
-        @test args.custom["listing"] !== nothing
-
-        # missing or invalid values produce a clean error
-        @test_throws ErrorException parse_args(["--jobs"])
-        @test_throws ErrorException parse_args(["--jobs=abc"])
-
-        # boolean flags reject values
-        @test_throws ErrorException parse_args(["--verbose=5"])
-    end
-
-    @testset "no arguments" begin
-        args = parse_args(String[])
-        @test args.jobs === nothing
-        @test args.verbose === nothing
-        @test args.quickfail === nothing
-        @test args.list === nothing
-        @test isempty(args.positionals)
-        @test isempty(args.custom)
-    end
-end
-
-@testset "filter_tests!" begin
-    @testset "empty positionals preserves all tests" begin
-        testsuite = Dict("a" => :(), "b" => :(), "c" => :())
-        args = parse_args(String[])
-        @test filter_tests!(testsuite, args) == true
-        @test length(testsuite) == 3
-    end
-
-    @testset "startswith matching" begin
-        testsuite = Dict("basic" => :(), "advanced" => :(), "basic_extra" => :())
-        args = parse_args(["basic"])
-        @test filter_tests!(testsuite, args) == false
-        @test haskey(testsuite, "basic")
-        @test haskey(testsuite, "basic_extra")
-        @test !haskey(testsuite, "advanced")
-    end
-
-    @testset "multiple positional filters" begin
-        testsuite = Dict("unit/a" => :(), "unit/b" => :(), "integration/c" => :(), "perf/d" => :())
-        args = parse_args(["unit", "integration"])
-        @test filter_tests!(testsuite, args) == false
-        @test haskey(testsuite, "unit/a")
-        @test haskey(testsuite, "unit/b")
-        @test haskey(testsuite, "integration/c")
-        @test !haskey(testsuite, "perf/d")
-    end
-
-    @testset "exclude tests basic" begin
-        testsuite = Dict("unit/a" => :(), "unit/b" => :(), "integration/c" => :(), "perf/d" => :())
-        args = parse_args(["!unit"])
-        @test filter_tests!(testsuite, args) == false
-        @test !haskey(testsuite, "unit/a")
-        @test !haskey(testsuite, "unit/b")
-        @test haskey(testsuite, "integration/c")
-        @test haskey(testsuite, "perf/d")
-    end
-
-    @testset "exclude included tests" begin
-        testsuite = Dict("unit/a" => :(), "unit/b" => :(), "integration/c" => :(), "perf/d" => :())
-        args = parse_args(["unit", "!unit/a"])
-        @test filter_tests!(testsuite, args) == false
-        @test !haskey(testsuite, "unit/a")
-        @test haskey(testsuite, "unit/b")
-        @test !haskey(testsuite, "integration/c")
-        @test !haskey(testsuite, "perf/d")
-    end
-
-    @testset "no matches yields empty suite" begin
-        testsuite = Dict("a" => :(), "b" => :())
-        args = parse_args(["nonexistent"])
-        @test filter_tests!(testsuite, args) == false
-        @test isempty(testsuite)
-    end
-
-    @testset "listing preserves all tests" begin
-        testsuite = Dict("a" => :(), "b" => :(), "c" => :())
-        args = parse_args(["--list"])
-        @test filter_tests!(testsuite, args) == false
-        @test length(testsuite) == 3
-    end
-
-    @testset "listing ignores positional filters" begin
-        testsuite = Dict("a" => :(), "b" => :())
-        args = parse_args(["--list", "a"])
-        @test filter_tests!(testsuite, args) == false
-        @test length(testsuite) == 2
-    end
-end
-
-@testset "find_tests edge cases" begin
-    @testset "empty directory" begin
-        mktempdir() do dir
-            @test isempty(find_tests(dir))
-        end
-    end
-
-    @testset "only runtests.jl" begin
-        mktempdir() do dir
-            write(joinpath(dir, "runtests.jl"), "@test true")
-            @test isempty(find_tests(dir))
-        end
-    end
-
-    @testset "nested subdirectories" begin
-        mktempdir() do dir
-            mkpath(joinpath(dir, "a", "b"))
-            write(joinpath(dir, "test1.jl"), "@test true")
-            write(joinpath(dir, "a", "test2.jl"), "@test true")
-            write(joinpath(dir, "a", "b", "test3.jl"), "@test true")
-            ts = find_tests(dir)
-            @test length(ts) == 3
-            @test haskey(ts, "test1")
-            @test haskey(ts, "a/test2")
-            @test haskey(ts, "a/b/test3")
-        end
-    end
-
-    @testset "non-.jl files ignored" begin
-        mktempdir() do dir
-            write(joinpath(dir, "test.jl"), "@test true")
-            write(joinpath(dir, "readme.md"), "# Readme")
-            write(joinpath(dir, "data.csv"), "1,2,3")
-            ts = find_tests(dir)
-            @test length(ts) == 1
-            @test haskey(ts, "test")
-        end
-    end
-end
-
-@testset "get_max_worker_rss" begin
-    rss = withenv("JULIA_TEST_MAXRSS_MB" => nothing) do
-        ParallelTestRunner.get_max_worker_rss()
-    end
-    @test rss > 0
-
-    rss = withenv("JULIA_TEST_MAXRSS_MB" => "1024") do
-        ParallelTestRunner.get_max_worker_rss()
-    end
-    @test rss == 1024 * 2^20
-end
-
-@testset "test_exe" begin
-    exe = ParallelTestRunner.test_exe(false)
-    @test any(contains("--color=no"), exe.exec)
-    @test any(contains("--project="), exe.exec)
-
-    exe = ParallelTestRunner.test_exe(true)
-    @test any(contains("--color=yes"), exe.exec)
-end
-
-@testset "truncate_line" begin
-    # short lines are untouched
-    @test ParallelTestRunner.truncate_line("short", 80) == "short"
-    @test ParallelTestRunner.truncate_line("x"^80, 80) == "x"^80
-
-    truncated = ParallelTestRunner.truncate_line("x"^100, 80)
-    @test length(truncated) == 80
-    @test endswith(truncated, "...")
-
-    # multi-byte characters must not break the cut: with byte indexing these
-    # would throw a StringIndexError when the cut lands mid-character
-    truncated = ParallelTestRunner.truncate_line("t" * "α"^100, 80)
-    @test length(truncated) == 80
-    @test endswith(truncated, "...")
-
-    truncated = ParallelTestRunner.truncate_line("€"^100, 40)
-    @test length(truncated) == 40
-    @test endswith(truncated, "...")
-end
-
-@testset "TestHistoryEntry" begin
-    flow = ParallelTestRunner.TestHistoryEntry(1,true)
-    fhigh = ParallelTestRunner.TestHistoryEntry(10,true)
-
-    slow = ParallelTestRunner.TestHistoryEntry(1,false)
-    shigh = ParallelTestRunner.TestHistoryEntry(10,false)
-
-    # irreflexive: lt(x, x) always yields false
-    @test !isless(flow, flow)
-    @test !isless(slow, slow)
-    @test !isless(fhigh, fhigh)
-    @test !isless(shigh, shigh)
-
-    # asymmetric: if lt(x, y) yields true then lt(y, x) yields false
-    @test isless(flow, fhigh)
-    @test !isless(fhigh, flow)
-    @test isless(shigh, flow)
-    @test !isless(flow, shigh)
-
-    # transitive: lt(x, y) && lt(y, z) implies lt(x, z)
-    @test isless(slow, shigh)
-    @test isless(shigh, flow)
-    @test isless(slow, flow)
-end
-
-# ── Integration tests ────────────────────────────────────────────────────────
-
-@testset "non-verbose mode" begin
-    testsuite = Dict("quiet" => :())
-    io = IOBuffer()
-    runtests(ParallelTestRunner, String[]; testsuite, stdout=io, stderr=io)
-    str = String(take!(io))
-    @test !contains(str, "started at")
-    @test !contains(str, "Available memory:")
-    @test !contains(str, "Max worker RSS:")
-    @test contains(str, "SUCCESS")
-end
-
-@testset "quickfail" begin
-    # To test that `--quickfail` does its job, we want to launch tests with one job, with
-    # the failing one launched as first: we should observe that the successful ones are
-    # never started.
-    testsuite = Dict(
-        "fail-test" => :( @test false ),
-        # If there's a bug in the launch order, having multiple successful tests increases
-        # the likelihood of hitting the bug, making investigation easier.
-        "pass-test1" => :( @test true ),
-        "pass-test2" => :( @test true ),
-        "pass-test3" => :( @test true ),
-        "pass-test4" => :( @test true ),
-        "pass-test5" => :( @test true ),
-    )
-    io = IOBuffer()
-    @test_throws Test.FallbackTestSetException begin
-        # Call `_runtests` so that we can enforce a run order.
-        ParallelTestRunner._runtests(
-            # Use a single job to make sure only `fail-test` is started.
-            ParallelTestRunner, parse_args(["--quickfail", "--verbose", "--jobs=1"]);
-            testsuite,
-            tests=["fail-test", "pass-test1", "pass-test2", "pass-test3", "pass-test4", "pass-test5"],
-            stdout=io,
-            stderr=io,
-        )
-    end
-    str = String(take!(io))
-    @test contains(str, r"fail-test .+ started at")
-    @test contains(str, r"fail-test .+ failed at")
-    @test contains(str, "FAILURE")
-    # `fail-test` is always launched first, `pass-test`s are never started.
-    @test !contains(str, r"pass-test[1-5] .+ started at")
-end
-
-@testset "positional filter end-to-end" begin
-    testsuite = Dict(
-        "unit/math" => :( @test 1 + 1 == 2 ),
-        "unit/string" => :( @test "a" * "b" == "ab" ),
-        "integration/api" => :( @test true ),
-    )
-    io = IOBuffer()
-    runtests(ParallelTestRunner, ["unit"]; testsuite, stdout=io, stderr=io)
-    str = String(take!(io))
-    @test contains(str, "Running 2 tests")
-    @test contains(str, "SUCCESS")
-end
-
-@testset "--list output" begin
-    # `--list` exits the process, so it has to run in a subprocess
-    mod = Module(:ListOutputTest)
-    history_file = ParallelTestRunner.get_history_file(mod)
-    function list_output()
-        code = """
-            using ParallelTestRunner
-            testsuite = Dict(name => :(@test true) for name in ("beta", "alpha", "gamma_longer"))
-            runtests(Module(:ListOutputTest), ["--list"]; testsuite)
-            """
-        readlines(`$(Base.julia_cmd()) --startup-file=no --project=$(Base.active_project()) --color=yes -e $code`)
-    end
-
-    rm(history_file; force=true)
-    @test list_output() == [
-            "Available tests:",
-            " - alpha",
-            " - beta",
-            " - gamma_longer"]
-
-    ParallelTestRunner.save_test_history(mod, (Dict("alpha" => 1.234, "gamma_longer" => 123.456), Set(["gamma_longer"])))
-    try
-        @test list_output() == [
-            "Available tests:",
-            " - alpha           (1.23s)",
-            " - beta",
-            "\e[31m × gamma_longer  (123.46s)\e[39m",
-        ]
-    finally
-        rm(history_file; force=true)
-    end
-end
-
-@testset "addworkers" begin
-    workers = addworkers(2)
-    @test length(workers) == 2
-    @test all(w -> w isa ParallelTestRunner.PTRWorker, workers)
-    @test all(w -> Base.process_running(w.w.proc), workers)
-    for w in workers
-        ParallelTestRunner.Malt.stop(w)
-    end
-    sleep(0.5)
-    @test all(w -> !Base.process_running(w.w.proc), workers)
-end
-
-@testset "multiple tests multiple jobs" begin
-    testsuite = Dict(
-        "m1" => :( @test 1 + 1 == 2 ),
-        "m2" => :( @test 2 + 2 == 4 ),
-        "m3" => :( @test 3 + 3 == 6 ),
-        "m4" => :( @test 4 + 4 == 8 ),
-    )
-    io = IOBuffer()
-    runtests(ParallelTestRunner, ["--jobs=2"]; testsuite, stdout=io, stderr=io)
-    str = String(take!(io))
-    @test contains(str, "Running 4 tests using 2 parallel jobs")
-    @test contains(str, "SUCCESS")
-end
-
-@testset "worker RSS recycling" begin
-    testsuite = Dict(
-        "alloc1" => :( @test true ),
-        "alloc2" => :( @test true ),
-        "alloc3" => :( @test true ),
-        "alloc4" => :( @test true ),
-    )
-    io = IOBuffer()
-    old_id_counter = ParallelTestRunner.ID_COUNTER[]
-    runtests(ParallelTestRunner, ["--jobs=1"]; testsuite, stdout=io, stderr=io, max_worker_rss=0)
-    str = String(take!(io))
-    @test contains(str, "SUCCESS")
-    @test ParallelTestRunner.ID_COUNTER[] == old_id_counter + length(testsuite)
-end
-
-@testset "mixed pass and fail" begin
-    testsuite = Dict(
-        "passes" => quote
-            @test true
-            @test 1 + 1 == 2
-        end,
-        "also_passes" => quote
-            @test true
-        end,
-        "fails" => quote
-            @test false
-        end,
-    )
-    io = IOBuffer()
-    @test_throws Test.FallbackTestSetException begin
-        runtests(ParallelTestRunner, String[]; testsuite, stdout=io, stderr=io)
-    end
-    str = String(take!(io))
-    @test contains(str, "FAILURE")
-    @test contains(str, "passes")
-    @test contains(str, "also_passes")
-    @test contains(str, "fails")
-end
-
-@testset "empty test suite" begin
-    testsuite = Dict{String,Expr}()
-    io = IOBuffer()
-    runtests(ParallelTestRunner, String[]; testsuite, stdout=io, stderr=io)
-    str = String(take!(io))
-    @test contains(str, "Running 0 tests")
-    @test contains(str, "SUCCESS")
-end
-
-# ── Serial tests ─────────────────────────────────────────────────────────────
-
-@testset "serial tests" begin
-    @testset "partition_tests" begin
-        @testset "basic partitioning preserves order" begin
-            tests = ["a", "b", "c", "d", "e"]
-            serial, parallel = ParallelTestRunner.partition_tests(tests, ["c", "a"])
-            @test serial == ["a", "c"]
-            @test parallel == ["b", "d", "e"]
-        end
-
-        @testset "empty serial list" begin
-            tests = ["x", "y", "z"]
-            serial, parallel = ParallelTestRunner.partition_tests(tests, String[])
-            @test isempty(serial)
-            @test parallel == tests
-        end
-
-        @testset "all tests serial" begin
-            tests = ["a", "b"]
-            serial, parallel = ParallelTestRunner.partition_tests(tests, ["a", "b"])
-            @test serial == ["a", "b"]
-            @test isempty(parallel)
-        end
-
-        @testset "unknown serial name throws" begin
-            tests = ["a", "b"]
-            @test_throws ArgumentError ParallelTestRunner.partition_tests(tests, ["a", "missing"])
-        end
-    end
-
-    @testset "serial tests run before parallel (default)" begin
-        serial_test_body = quote
-            children = _count_child_pids($(getpid()))
-            # Make sure serial tests run alone.
-            if children >= 0
-                @test children == 1
-            end
-        end
-        testsuite = Dict(
-            "serial_1" => serial_test_body,
-            "serial_2" => serial_test_body,
-            "serial_3" => serial_test_body,
-            "parallel_1" => :(),
-            "parallel_2" => :(),
-            "parallel_3" => :(),
-        )
-        io = IOBuffer()
-        jobs = 2
-        old_id_counter = ParallelTestRunner.ID_COUNTER[]
-        @show_if_error io runtests(ParallelTestRunner, ["--jobs=$(jobs)", "--verbose"];
-                                   testsuite, stdout=io, stderr=io,
-                                   init_code=:(include($(joinpath(@__DIR__, "utils.jl")))),
-                                   serial=["serial_1", "serial_2", "serial_3"])
-        str = String(take!(io))
-        @test contains(str, "Running 6 tests using 2 parallel jobs")
-        @test contains(str, "3 serial test(s) will run before")
-        @test contains(str, "SUCCESS")
-        @test ParallelTestRunner.ID_COUNTER[] == old_id_counter + jobs
-    end
-
-    @testset "serial tests run after parallel" begin
-        serial_test_body = quote
-            children = _count_child_pids($(getpid()))
-            # Make sure serial tests run alone.
-            if children >= 0
-                @test children == 1
-            end
-        end
-        testsuite = Dict(
-            "serial_1" => serial_test_body,
-            "serial_2" => serial_test_body,
-            "serial_3" => serial_test_body,
-            "parallel_1" => :(),
-            "parallel_2" => :(),
-            "parallel_3" => :(),
-        )
-        io = IOBuffer()
-        ioc = IOContext(io, :color => true)
-        old_id_counter = ParallelTestRunner.ID_COUNTER[]
-        @show_if_error io runtests(ParallelTestRunner, ["--jobs=2", "--verbose"];
-                                   testsuite, stdout=ioc, stderr=ioc,
-                                   init_code=:(include($(joinpath(@__DIR__, "utils.jl")))),
-                                   serial=["serial_1", "serial_2", "serial_3"], serial_position=:after)
-        str = String(take!(io))
-        @test contains(str, "Running 6 tests using 2 parallel jobs")
-        @test contains(str, "3 serial test(s) will run after")
-        @test contains(str, "SUCCESS")
-        @test ParallelTestRunner.ID_COUNTER[] == old_id_counter + 2
-    end
-
-    @testset "serial_position validation" begin
-        testsuite = Dict("a" => :())
-        io = IOBuffer()
-        @test_throws ArgumentError runtests(ParallelTestRunner, String[];
-                                            testsuite, stdout=io, stderr=io,
-                                            serial_position=:middle)
-    end
-
-    @testset "all tests serial" begin
-        testsuite = Dict(
-            "a" => :(),
-            "b" => :(),
-            "c" => :(),
-            "d" => :(),
-        )
-        io = IOBuffer()
-        old_id_counter = ParallelTestRunner.ID_COUNTER[]
-        runtests(ParallelTestRunner, ["--jobs=3", "--verbose"];
-                 testsuite, stdout=io, stderr=io,
-                 serial=["a", "b", "c", "d"])
-        str = String(take!(io))
-        @test contains(str, "Running 4 tests using 1 parallel jobs")
-        @test contains(str, "4 serial test(s) will run before")
-        @test contains(str, "SUCCESS")
-        @test ParallelTestRunner.ID_COUNTER[] == old_id_counter + 1
-    end
-
-    @testset "empty serial list is a no-op" begin
-        testsuite = Dict(
-            "a" => :(),
-            "b" => :(),
-        )
-        io = IOBuffer()
-        runtests(ParallelTestRunner, ["--jobs=2"]; testsuite, stdout=io, stderr=io,
-                 serial=String[])
-        str = String(take!(io))
-        @test !contains(str, "serial")
-        @test contains(str, "SUCCESS")
-    end
-
-    @testset "parallel tests less than requested jobs" begin
-        testsuite = Dict(
-            "s1" => :(),
-            "s2" => :(),
-            "p1" => :(),
-            "p2" => :(),
-        );
-        io = IOBuffer()
-        old_id_counter = ParallelTestRunner.ID_COUNTER[]
-        runtests(ParallelTestRunner, ["--jobs=3"]; testsuite, stdout=io, stderr=io,
-                 serial=["s1", "s2"])
-        str = String(take!(io))
-        # We have 4 total tests, requested 3 jobs, but only 2 tests are run in parallel, so
-        # 2 is the maximum parallelism we expect, and the number of new workers we spawn.
-        @test contains(str, "Running 4 tests using 2 parallel jobs")
-        @test contains(str, "2 serial test(s)")
-        @test contains(str, "SUCCESS")
-        @test ParallelTestRunner.ID_COUNTER[] == old_id_counter + 2
-    end
-
-    @testset "serial names filtered by positional args" begin
-        testsuite = Dict(
-            "unit/a" => :(),
-            "unit/b" => :(),
-            # This test file shouldn't called, we use `@test false` to make sure it's not.
-            "integration/c" => :(@test false),
-        )
-        io = IOBuffer()
-        runtests(ParallelTestRunner, ["unit"]; testsuite, stdout=io, stderr=io,
-                 serial=["unit/a", "integration/c"])
-        str = String(take!(io))
-        @test contains(str, "Running 2 tests")
-        @test contains(str, "1 serial test(s)")
-        @test contains(str, "SUCCESS")
-    end
-
-    @testset "crashing serial test" begin
-        serial_test_body = quote
-            children = _count_child_pids($(getpid()))
-            # Make sure serial tests run alone.
-            if children >= 0
-                @test children == 1
-            end
-        end
-
-        testsuite = Dict(
-            "s1" => serial_test_body,
-            "s2" => serial_test_body,
-            "s3" => serial_test_body,
-            "s4" => :(ccall(:abort, Nothing, ())),
-            "p1" => :(),
-            "p2" => :(),
-        )
-        io = IOBuffer()
-        ioc = IOContext(io, :color => true)
-        old_id_counter = ParallelTestRunner.ID_COUNTER[]
-        jobs = 2
-        @test_throws Test.FallbackTestSetException("Test run finished with errors") begin
-            runtests(ParallelTestRunner, ["--jobs=$(jobs)", "--verbose"];
-                     testsuite, stdout=ioc, stderr=ioc,
-                     init_code=:(include($(joinpath(@__DIR__, "utils.jl")))),
-                     serial=["s1", "s2", "s3", "s4"])
-        end
-        str = String(take!(io))
-        @test contains(str, "Running 6 tests using 2 parallel jobs")
-        @test contains(str, "4 serial test(s)")
-        @test contains(str, "FAILURE")
-        # We'll use jobs + 1 workers because one will crash.
-        @test ParallelTestRunner.ID_COUNTER[] == old_id_counter + jobs + 1
-    end
-
-    @testset "quickfail in serial phase before parallel" begin
-        # The failing serial test runs first: the remaining serial tests and the whole
-        # parallel batch should never be started.
-        testsuite = Dict(
-            "fail-serial" => :( @test false ),
-            "pass-serial1" => :( @test true ),
-            "pass-serial2" => :( @test true ),
-            "pass-parallel1" => :( @test true ),
-            "pass-parallel2" => :( @test true ),
-            "pass-parallel3" => :( @test true ),
-        )
-        io = IOBuffer()
-        @test_throws Test.FallbackTestSetException begin
-            # Call `_runtests` so that we can enforce a run order.
-            ParallelTestRunner._runtests(
-                ParallelTestRunner, parse_args(["--quickfail", "--verbose", "--jobs=2"]);
-                testsuite,
-                tests=["fail-serial", "pass-serial1", "pass-serial2",
-                       "pass-parallel1", "pass-parallel2", "pass-parallel3"],
-                serial=["fail-serial", "pass-serial1", "pass-serial2"],
-                stdout=io,
-                stderr=io,
-            )
-        end
-        str = String(take!(io))
-        @test contains(str, "3 serial test(s) will run before")
-        @test contains(str, r"fail-serial .+ started at")
-        @test contains(str, r"fail-serial .+ failed at")
-        @test contains(str, "FAILURE")
-        # `fail-serial` is always launched first, the other tests are never started.
-        @test !contains(str, r"pass-serial[12] .+ started at")
-        @test !contains(str, r"pass-parallel[1-3] .+ started at")
-    end
-
-    @testset "quickfail in parallel phase with serial tests after" begin
-        # The failing parallel test runs first: the remaining parallel tests and the
-        # whole serial batch should never be started.
-        testsuite = Dict(
-            "fail-parallel" => :( @test false ),
-            "pass-parallel1" => :( @test true ),
-            "pass-parallel2" => :( @test true ),
-            "pass-serial1" => :( @test true ),
-            "pass-serial2" => :( @test true ),
-        )
-        io = IOBuffer()
-        @test_throws Test.FallbackTestSetException begin
-            # Call `_runtests` so that we can enforce a run order.
-            ParallelTestRunner._runtests(
-                # Use a single job to make sure only `fail-parallel` is started.
-                ParallelTestRunner, parse_args(["--quickfail", "--verbose", "--jobs=1"]);
-                testsuite,
-                tests=["fail-parallel", "pass-parallel1", "pass-parallel2",
-                       "pass-serial1", "pass-serial2"],
-                serial=["pass-serial1", "pass-serial2"],
-                serial_position=:after,
-                stdout=io,
-                stderr=io,
-            )
-        end
-        str = String(take!(io))
-        @test contains(str, "2 serial test(s) will run after")
-        @test contains(str, r"fail-parallel .+ started at")
-        @test contains(str, r"fail-parallel .+ failed at")
-        @test contains(str, "FAILURE")
-        # `fail-parallel` is always launched first, the other tests are never started.
-        @test !contains(str, r"pass-parallel[12] .+ started at")
-        @test !contains(str, r"pass-serial[12] .+ started at")
-    end
-
-    @testset "quickfail in serial phase after parallel" begin
-        # All parallel tests pass, then the failing serial test runs first in the serial
-        # phase: the remaining serial tests should never be started.
-        testsuite = Dict(
-            "pass-parallel1" => :( @test true ),
-            "pass-parallel2" => :( @test true ),
-            "fail-serial" => :( @test false ),
-            "pass-serial1" => :( @test true ),
-            "pass-serial2" => :( @test true ),
-        )
-        io = IOBuffer()
-        @test_throws Test.FallbackTestSetException begin
-            # Call `_runtests` so that we can enforce a run order.
-            ParallelTestRunner._runtests(
-                ParallelTestRunner, parse_args(["--quickfail", "--verbose", "--jobs=2"]);
-                testsuite,
-                tests=["pass-parallel1", "pass-parallel2",
-                       "fail-serial", "pass-serial1", "pass-serial2"],
-                serial=["fail-serial", "pass-serial1", "pass-serial2"],
-                serial_position=:after,
-                stdout=io,
-                stderr=io,
-            )
-        end
-        str = String(take!(io))
-        @test contains(str, "3 serial test(s) will run after")
-        # The parallel batch runs to completion before the serial phase starts.
-        @test contains(str, r"pass-parallel1 .+ started at")
-        @test contains(str, r"pass-parallel2 .+ started at")
-        @test contains(str, r"fail-serial .+ started at")
-        @test contains(str, r"fail-serial .+ failed at")
-        @test contains(str, "FAILURE")
-        # The serial tests after `fail-serial` are never started.
-        @test !contains(str, r"pass-serial[12] .+ started at")
-    end
-
-    @testset "run previously failed tests first" begin
-        # a previously-failed test must start before previously-passing tests,
-        # even if it has a much shorter historical duration than they do
-        mod = @eval(Main, module $(gensym(:failfirstserial)) end)
-
-        testsuite = Dict(
-            "long-serial-pass" => :(@test true),
-            "mid-omitted-serial-pass" => :(@test true),
-            "short-serial-fail" => :(@test true),
-            "long-pass" => :(@test true),
-            "mid-omitted-pass" => :(@test true),
-            "mid-pass" => :(@test true),
-            "short-fail" => :(@test true),
-        )
-        serial = filter(collect(keys(testsuite))) do k
-            contains(k, "serial")
-        end
-
-        ParallelTestRunner.save_test_history(mod, (Dict(
-            "long-serial-pass" => 10.0,
-            "short-serial-fail" => 1.0,
-            "long-pass" => 10.0,
-            "mid-pass" => 5.0,
-            "short-fail" => 1.0,
-        ), Set(["short-serial-fail", "short-fail"])))
-
-        io = IOBuffer()
-        runtests(
-            mod, parse_args(["--jobs=1", "--verbose"]);
-            testsuite,
-            stdout=io,
-            stderr=io,
-            serial
-        )
-
-        str = String(take!(io))
-        @test contains(str, "SUCCESS")
-
-        # create a mapping of test names to the character offset of their start times (lower is earlier)
-        started_at = Dict(
-            name => (m = match(Regex("$(name) .+ started at"), str); @test m !== nothing; m === nothing ? typemax(Int) : m.offset)
-            for name in keys(testsuite)
-        )
-
-        # normal
-        @test started_at["short-fail"] < started_at["mid-pass"]
-        @test started_at["short-fail"] < started_at["long-pass"]
-        @test started_at["short-fail"] < started_at["mid-omitted-pass"]
-        # among the remaining (previously-passing) tests, longer ones still run first
-        @test started_at["long-pass"] < started_at["mid-pass"]
-
-        # serial
-        @test started_at["short-serial-fail"] < started_at["long-serial-pass"]
-        @test started_at["short-serial-fail"] < started_at["mid-omitted-serial-pass"]
-    end
-end
-
-@testset "recycle_on_failure" begin
-    # Call `_runtests` throughout, so that we can enforce a run order, and use a single job,
-    # so that all tests share the same pool slot: a test only gets a new worker if the
-    # previous one was recycled.
-    testsuite = Dict(
-        "fail1" => :( @test false ),
-        "pass1" => :( @test true ),
-        "fail2" => :( @test false ),
-        "pass2" => :( @test true ),
-    )
-    tests = ["fail1", "pass1", "fail2", "pass2"]
-
-    @testset "workers are reused across failures by default" begin
-        io = IOBuffer()
-        old_id_counter = ParallelTestRunner.ID_COUNTER[]
-        @test_throws Test.FallbackTestSetException begin
-            ParallelTestRunner._runtests(
-                ParallelTestRunner, parse_args(["--jobs=1"]);
-                testsuite,
-                tests,
-                stdout=io,
-                stderr=io,
-            )
-        end
-        str = String(take!(io))
-        @test contains(str, "FAILURE")
-        # A failing test does not recycle its worker, so a single one runs all four tests.
-        @test ParallelTestRunner.ID_COUNTER[] == old_id_counter + 1
-    end
-
-    @testset "worker is recycled after a failed test" begin
-        io = IOBuffer()
-        old_id_counter = ParallelTestRunner.ID_COUNTER[]
-        @test_throws Test.FallbackTestSetException begin
-            ParallelTestRunner._runtests(
-                ParallelTestRunner, parse_args(["--jobs=1"]);
-                testsuite,
-                tests,
-                stdout=io,
-                stderr=io,
-                recycle_on_failure=true,
-            )
-        end
-        str = String(take!(io))
-        @test contains(str, "FAILURE")
-        # `fail1` and `fail2` recycle their worker, so `pass1` and `pass2` each need a fresh
-        # one: 1 initial worker + 2 replacements.
-        @test ParallelTestRunner.ID_COUNTER[] == old_id_counter + 3
-    end
-end
-
-@testset "retries" begin
-    # A test that fails on its first attempt and passes on any subsequent one, by recording
-    # attempts in a file: the worker running the retry is a different process, so the marker
-    # has to live outside of it.
-    flaky_test(marker, body=:( @test true )) = quote
-        if isfile($marker)
-            $body
-        else
-            touch($marker)
-            @test false
-        end
-    end
-
-    @testset "failed test passing on retry is reported as passing" begin
-        mktempdir() do dir
-            testsuite = Dict(
-                "flaky" => flaky_test(joinpath(dir, "flaky")),
-                "passes" => :( @test true ),
-            )
-            io = IOBuffer()
-            @show_if_error io ParallelTestRunner._runtests(
-                ParallelTestRunner, parse_args(["--jobs=1"]);
-                testsuite,
-                tests=["flaky", "passes"],
-                stdout=io,
-                stderr=io,
-                retries=1,
-            )
-            str = String(take!(io))
-            # Only the failed test is retried, and its retried result is the one reported.
-            @test contains(str, "Retrying 1 failed test")
-            @test contains(str, "SUCCESS")
-            # Two results in total: the failed attempt of `flaky` was replaced by the
-            # retried one, rather than reported next to it.
-            @test contains(str, r"Overall +\| +2 +2 ")
-        end
-    end
-
-    @testset "persistent failure is retried and reported once" begin
-        testsuite = Dict(
-            "always_fails" => :( @test false ),
-            "passes" => :( @test true ),
-        )
-        io = IOBuffer()
-        @test_throws Test.FallbackTestSetException begin
-            ParallelTestRunner._runtests(
-                ParallelTestRunner, parse_args(["--jobs=1"]);
-                testsuite,
-                tests=["always_fails", "passes"],
-                stdout=io,
-                stderr=io,
-                retries=2,
-            )
-        end
-        str = String(take!(io))
-        @test contains(str, "FAILURE")
-        # Both retry rounds run, and each of them fails again.
-        @test length(collect(eachmatch(r"always_fails.*failed", str))) == 3
-        # Despite the three attempts, the test is reported exactly once, as a failure.
-        @test contains(str, r"always_fails +\| +1 +1 ")
-    end
-
-    # `serial_position=:after` returns the live serial worker to the pool immediately
-    # before the retry phase, so it is the configuration where the "alone" invariant is
-    # easiest to break.
-    @testset "retried test runs alone (serial=$serial, $serial_position)" for
-            (serial, serial_position) in ((String[], :before),
-                                          (["pass3"], :before),
-                                          (["pass3"], :after))
-        mktempdir() do dir
-            # On its retry, the flaky test checks it is the only worker left alive.
-            check_alone = quote
-                children = _count_child_pids($(getpid()))
-                if children >= 0
-                    @test children == 1
-                end
-            end
-            testsuite = Dict(
-                "flaky" => flaky_test(joinpath(dir, "flaky"), check_alone),
-                "pass1" => :( @test true ),
-                "pass2" => :( @test true ),
-                "pass3" => :( @test true ),
-            )
-            io = IOBuffer()
-            @show_if_error io ParallelTestRunner._runtests(
-                ParallelTestRunner, parse_args(["--jobs=3"]);
-                testsuite,
-                tests=["flaky", "pass1", "pass2", "pass3"],
-                init_code=:(include($(joinpath(@__DIR__, "utils.jl")))),
-                serial,
-                serial_position,
-                stdout=io,
-                stderr=io,
-                retries=1,
-            )
-            str = String(take!(io))
-            @test length(collect(eachmatch(r"failed", str))) == 2
-            @test contains(str, "SUCCESS")
-        end
-    end
-
-    @testset "failing retry does not reuse its worker" begin
-        testsuite = Dict(
-            "failA" => :( @test false ),
-            "failB" => :( @test false ),
-        )
-        io = IOBuffer()
-        @test_throws Test.FallbackTestSetException begin
-            ParallelTestRunner._runtests(
-                ParallelTestRunner, parse_args(["--jobs=1"]);
-                testsuite,
-                tests=["failA", "failB"],
-                stdout=io,
-                stderr=io,
-                retries=1,
-            )
-        end
-        str = String(take!(io))
-        main, retry = split(str, "Retrying")
-        ids(s) = [m[1] for m in eachmatch(r"fail[AB] +\((\d+)\)", s)]
-        # the main run reuses one worker: `recycle_on_failure` is off by default
-        @test length(ids(main)) == 2 && allequal(ids(main))
-        # the retry round recycles after every non-pass, so each test gets its own worker
-        @test length(ids(retry)) == 2 && allunique(ids(retry))
-    end
-
-    @testset "no retries by default" begin
-        mktempdir() do dir
-            testsuite = Dict("flaky" => flaky_test(joinpath(dir, "flaky")))
-            io = IOBuffer()
-            @test_throws Test.FallbackTestSetException begin
-                ParallelTestRunner._runtests(
-                    ParallelTestRunner, parse_args(["--jobs=1"]);
-                    testsuite,
-                    tests=["flaky"],
-                    stdout=io,
-                    stderr=io,
-                )
-            end
-            str = String(take!(io))
-            @test !contains(str, "Retrying")
-            @test contains(str, "FAILURE")
-        end
-    end
-
-    @testset "quickfail skips retries" begin
-        mktempdir() do dir
-            testsuite = Dict(
-                "flaky" => flaky_test(joinpath(dir, "flaky")),
-                "passes" => :( @test true ),
-            )
-            io = IOBuffer()
-            @test_throws Test.FallbackTestSetException begin
-                ParallelTestRunner._runtests(
-                    ParallelTestRunner, parse_args(["--quickfail", "--jobs=1"]);
-                    testsuite,
-                    tests=["flaky", "passes"],
-                    stdout=io,
-                    stderr=io,
-                    retries=1,
-                )
-            end
-            str = String(take!(io))
-            # The run stopped early on purpose, retrying would defeat that.
-            @test !contains(str, "Retrying")
-            @test contains(str, "FAILURE")
-        end
-    end
-end
-
-# This testset should always be the last one, don't add anything after this.
-# We want to make sure there are no running workers at the end of the tests.
+# All workers must have been stopped once `runtests` returns.
 @testset "no workers running" begin
     children = _count_child_pids()
     if children >= 0
         @test children == 0
     end
-end
-
 end
